增加重置sessionToken,简化SaveModule,适配Java 8

+++ PhigrosLibrary/src/main/java/given/phigros/ByteReader.java
@@ -2,7 +2,7 @@ package given.phigros;
 
 class ByteReader {
     byte[] data;
-    int position;
+    short position;
 
     ByteReader(byte[] data) {
         this.data = data;
